Move functions and templates to appropriate sub directories
