+++ util/speak.py
